fix: fixes movimentation page widgets
