+++ lib/details/useCase/get_chart_info_usecase.dart
@@ -1,7 +1,7 @@
 import '../../shared/apiCoingecko/response/chartInfo/chart_info_mapper.dart';
 import '../../shared/apiCoingecko/response/chartInfo/chart_info_response.dart';
-import '../repository/chart_info_repo.dart';
 import '../models/chart_info_model.dart';
+import '../repository/chart_info_repo.dart';
 
 class GetChartInfoUsecase {
     final ChartInfoRepo repo;
