--- lib/movimentations/widgets/divider_bottom_sheet_detail.dart
@@ -1,15 +1,0 @@
-import 'package:flutter/material.dart';
-
-class DividerBottomSheetDetail extends StatelessWidget {
-  const DividerBottomSheetDetail({
-    Key? key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return const Divider(
-      thickness: 1.3,
-      height: 22,
-    );
-  }
-}
