+++ lib/portfolio/widgets/currency_list_card.dart
@@ -3,8 +3,8 @@ import 'package:decimal/intl.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:intl/intl.dart';
-import '../../details/view/details_page.dart';
 
+import '../../details/view/details_page.dart';
 import '../../shared/args/arguments.dart';
 import '../../shared/styles/colors.dart';
 import '../models/crypto_model.dart';
