+++ lib/shared/routes/routes.dart
@@ -1,11 +1,9 @@
-import '../../review/view/review_page.dart';
-import '../../success/view/success_page.dart';
-
 import '../../converter/view/converter_page.dart';
-
 import '../../details/view/details_page.dart';
 import '../../movimentations/view/movimentations_page.dart';
 import '../../portfolio/view/crypto_page.dart';
+import '../../review/view/review_page.dart';
+import '../../success/view/success_page.dart';
 
 var appRoutes = {
   CryptoPage.route: (context) => const CryptoPage(),
